refactor: Folder restructuring
