--- public/storage
@@ -1,1 +1,0 @@
-/home/ecanquiz/devenv/warehouse/static-server-node/storage
